# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  apierror.ErrorResponse:
    properties:
      error:
        description: Stable error code (e.g., "invalid_credentials")
        type: string
      message:
        description: Human-readable message
        type: string
    type: object
...
  handler.artifactFullResponse:
    properties:
      content:
        type: object
      created_at:
        type: string
      id:
        type: string
      iteration:
        type: integer
      kind:
        type: string
      parent_id:
        type: string
      producer_agent:
        type: string
      status:
        type: string
      summary:
        type: string
      task_id:
        type: string
    type: object
...
  handler.worktreeResponse:
    properties:
      allocated_at:
        type: string
      base_branch:
        type: string
      branch_name:
        type: string
      id:
        type: string
      released_at:
        type: string
      state:
        type: string
      subtask_id:
        type: string
      task_id:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: Backend API с авторизацией на JWT токенах
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Backend API
  version: "1.0"
paths:
...
  /tasks/{id}/artifacts:
    get:
      description: Все артефакты задачи в порядке создания. Content не загружается;
        для полного — GET /tasks/{id}/artifacts/{artifactId}.
      parameters:
      - description: Task UUID
        in: path
        name: id
        required: true
        type: string
      - description: только status='ready' (default false — включая superseded)
        in: query
        name: only_ready
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apierror.ErrorResponse'
        "403":
          description: no access to task
          schema:
            $ref: '#/definitions/apierror.ErrorResponse'
        "404":
          description: task not found
          schema:
            $ref: '#/definitions/apierror.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List task artifacts (metadata)
      tags:
      - orchestration-v2
  /tasks/{id}/artifacts/{artifactId}:
    get:
      parameters:
      - description: Task UUID
        in: path
        name: id
        required: true
        type: string
      - description: Artifact UUID
        in: path
        name: artifactId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.artifactFullResponse'
        "403":
          description: no access to task
          schema:
            $ref: '#/definitions/apierror.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apierror.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get artifact by ID (with content)
      tags:
      - orchestration-v2
...
  /tasks/{id}/router-decisions:
    get:
      description: Все решения Router'а в порядке step_no. encrypted_raw_response
        НЕ возвращается.
      parameters:
      - description: Task UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apierror.ErrorResponse'
        "403":
          description: no access to task
          schema:
            $ref: '#/definitions/apierror.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apierror.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List router decisions (timeline)
      tags:
      - orchestration-v2
...
  /worktrees:
    get:
      description: Список git worktree'ев. Без task_id — admin-only; с task_id — нужен
        доступ к задаче.
      parameters:
      - description: Filter by task UUID
        in: query
        name: task_id
        type: string
      - description: Filter by state (allocated|in_use|released)
        in: query
        name: state
        type: string
      - description: Max items (default 200, capped at 200)
        in: query
        name: limit
        type: integer
      - description: Offset (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apierror.ErrorResponse'
        "403":
          description: no task_id requires admin role; task_id requires task access
          schema:
            $ref: '#/definitions/apierror.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List worktrees (debug)
      tags:
      - orchestration-v2
  /worktrees/{id}/release:
    post:
      description: Принудительно выполняет git worktree remove --force и помечает
        запись released. Admin-only. Idempotent NOT — повторный вызов на released
        worktree возвращает 409.
      parameters:
      - description: Worktree UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.worktreeResponse'
        "400":
          description: invalid uuid
          schema:
            $ref: '#/definitions/apierror.ErrorResponse'
        "403":
          description: admin-only
          schema:
            $ref: '#/definitions/apierror.ErrorResponse'
        "404":
          description: worktree not found
          schema:
            $ref: '#/definitions/apierror.ErrorResponse'
        "409":
          description: worktree already released
          schema:
            $ref: '#/definitions/apierror.ErrorResponse'
        "503":
          description: worktree manager not configured
          schema:
            $ref: '#/definitions/apierror.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Manually release a worktree (admin)
      tags:
      - orchestration-v2
securityDefinitions:
  ApiKeyAuth:
    description: 'Long-lived API key for programmatic access. Format: wibe_<key>'
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
  OAuth2Password:
    flow: password
    tokenUrl: /api/v1/auth/login
    type: oauth2
swagger: "2.0"